Исправляет использование устаревшего синтаксиса, активно использует оператор spread, убирает вспомогательные абстактные функции в utils
